refactor: split large files for navigability post-v2.0.0
api.go (2,630→129), main.go (1,009→385), updater.go (1,149→593),
bolt.go (747→457). Pure code moves within same packages — no
behaviour changes, all tests pass unchanged.
