+++ internal/store/bolt.go
@@ -8,9 +8,6 @@ import (
 	"time"
 
 	bolt "go.etcd.io/bbolt"
-
-	"github.com/Will-Luck/Docker-Sentinel/internal/notify"
-	"github.com/Will-Luck/Docker-Sentinel/internal/registry"
 )
 
 var (
@@ -435,219 +432,6 @@ func (s *Store) GetAllSettings() (map[string]string, error) {
 	return result, err
 }
 
-// NotificationConfig represents persisted notification settings.
-type NotificationConfig struct {
-	GotifyURL      string            `json:"gotify_url"`
-	GotifyToken    string            `json:"gotify_token"`
-	WebhookURL     string            `json:"webhook_url"`
-	WebhookHeaders map[string]string `json:"webhook_headers"`
-}
-
-// GetNotificationConfig loads the notification configuration from the settings bucket.
-func (s *Store) GetNotificationConfig() (NotificationConfig, error) {
-	var cfg NotificationConfig
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketSettings)
-		v := b.Get([]byte("notification_config"))
-		if v == nil {
-			return nil
-		}
-		return json.Unmarshal(v, &cfg)
-	})
-	return cfg, err
-}
-
-// SetNotificationConfig saves the notification configuration to the settings bucket.
-func (s *Store) SetNotificationConfig(cfg NotificationConfig) error {
-	data, err := json.Marshal(cfg)
-	if err != nil {
-		return fmt.Errorf("marshal notification config: %w", err)
-	}
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketSettings)
-		return b.Put([]byte("notification_config"), data)
-	})
-}
-
-// GetNotificationChannels loads notification channels from the settings bucket.
-// Handles legacy format migration: if stored data is a JSON object (old NotificationConfig),
-// it converts it to the new channel array format.
-func (s *Store) GetNotificationChannels() ([]notify.Channel, error) {
-	var channels []notify.Channel
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketSettings)
-		v := b.Get([]byte("notification_channels"))
-		if v == nil {
-			// Try legacy key migration.
-			legacy := b.Get([]byte("notification_config"))
-			if legacy != nil {
-				channels = migrateFromLegacy(legacy)
-			}
-			return nil
-		}
-		return json.Unmarshal(v, &channels)
-	})
-	return channels, err
-}
-
-// SetNotificationChannels saves notification channels to the settings bucket.
-func (s *Store) SetNotificationChannels(channels []notify.Channel) error {
-	data, err := json.Marshal(channels)
-	if err != nil {
-		return fmt.Errorf("marshal notification channels: %w", err)
-	}
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketSettings)
-		return b.Put([]byte("notification_channels"), data)
-	})
-}
-
-// migrateFromLegacy converts old NotificationConfig JSON into Channel entries.
-func migrateFromLegacy(data []byte) []notify.Channel {
-	var legacy NotificationConfig
-	if err := json.Unmarshal(data, &legacy); err != nil {
-		return nil
-	}
-
-	var channels []notify.Channel
-
-	if legacy.GotifyURL != "" {
-		settings, _ := json.Marshal(notify.GotifySettings{
-			URL:   legacy.GotifyURL,
-			Token: legacy.GotifyToken,
-		})
-		channels = append(channels, notify.Channel{
-			ID:       notify.GenerateID(),
-			Type:     notify.ProviderGotify,
-			Name:     "Gotify",
-			Enabled:  true,
-			Settings: settings,
-		})
-	}
-
-	if legacy.WebhookURL != "" {
-		settings, _ := json.Marshal(notify.WebhookSettings{
-			URL:     legacy.WebhookURL,
-			Headers: legacy.WebhookHeaders,
-		})
-		channels = append(channels, notify.Channel{
-			ID:       notify.GenerateID(),
-			Type:     notify.ProviderWebhook,
-			Name:     "Webhook",
-			Enabled:  true,
-			Settings: settings,
-		})
-	}
-
-	return channels
-}
-
-// AddIgnoredVersion records that a specific version should be ignored for a container.
-// The value stored under each container name is a JSON array of version strings.
-func (s *Store) AddIgnoredVersion(containerName, version string) error {
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketIgnoredVersions)
-
-		existing := b.Get([]byte(containerName))
-		var versions []string
-		if existing != nil {
-			if err := json.Unmarshal(existing, &versions); err != nil {
-				return fmt.Errorf("unmarshal ignored versions: %w", err)
-			}
-		}
-
-		// Skip if already present.
-		for _, v := range versions {
-			if v == version {
-				return nil
-			}
-		}
-
-		versions = append(versions, version)
-		data, err := json.Marshal(versions)
-		if err != nil {
-			return fmt.Errorf("marshal ignored versions: %w", err)
-		}
-		return b.Put([]byte(containerName), data)
-	})
-}
-
-// GetIgnoredVersions returns all ignored versions for a container.
-// Returns an empty slice if none are stored.
-func (s *Store) GetIgnoredVersions(containerName string) ([]string, error) {
-	var versions []string
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketIgnoredVersions)
-		v := b.Get([]byte(containerName))
-		if v == nil {
-			return nil
-		}
-		return json.Unmarshal(v, &versions)
-	})
-	if versions == nil {
-		versions = []string{}
-	}
-	return versions, err
-}
-
-// ClearIgnoredVersions removes all ignored versions for a container.
-func (s *Store) ClearIgnoredVersions(containerName string) error {
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketIgnoredVersions)
-		return b.Delete([]byte(containerName))
-	})
-}
-
-// GetRegistryCredentials loads registry credentials from the registry_credentials bucket.
-func (s *Store) GetRegistryCredentials() ([]registry.RegistryCredential, error) {
-	var creds []registry.RegistryCredential
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketRegistryCreds)
-		v := b.Get([]byte("credentials"))
-		if v == nil {
-			return nil
-		}
-		return json.Unmarshal(v, &creds)
-	})
-	return creds, err
-}
-
-// SetRegistryCredentials saves registry credentials to the registry_credentials bucket.
-func (s *Store) SetRegistryCredentials(creds []registry.RegistryCredential) error {
-	data, err := json.Marshal(creds)
-	if err != nil {
-		return fmt.Errorf("marshal registry credentials: %w", err)
-	}
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketRegistryCreds)
-		return b.Put([]byte("credentials"), data)
-	})
-}
-
-// SaveRateLimits persists rate limit state for all registries.
-func (s *Store) SaveRateLimits(data []byte) error {
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketRateLimits)
-		return b.Put([]byte("state"), data)
-	})
-}
-
-// LoadRateLimits loads persisted rate limit state.
-// Returns nil, nil if nothing is stored.
-func (s *Store) LoadRateLimits() ([]byte, error) {
-	var data []byte
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketRateLimits)
-		v := b.Get([]byte("state"))
-		if v != nil {
-			data = make([]byte, len(v))
-			copy(data, v)
-		}
-		return nil
-	})
-	return data, err
-}
-
 // CountHistory returns the number of entries in the history bucket.
 // Uses bucket stats for O(1) counting.
 func (s *Store) CountHistory() (int, error) {
@@ -670,78 +454,4 @@ func (s *Store) CountSnapshots() (int, error) {
 		return nil
 	})
 	return count, err
-}
-
-// SaveGHCRCache persists GHCR alternative detection cache.
-func (s *Store) SaveGHCRCache(data []byte) error {
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketGHCRAlternatives)
-		return b.Put([]byte("cache"), data)
-	})
-}
-
-// LoadGHCRCache loads persisted GHCR alternative cache.
-// Returns nil, nil if nothing is stored.
-func (s *Store) LoadGHCRCache() ([]byte, error) {
-	var data []byte
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketGHCRAlternatives)
-		v := b.Get([]byte("cache"))
-		if v != nil {
-			data = make([]byte, len(v))
-			copy(data, v)
-		}
-		return nil
-	})
-	return data, err
-}
-
-// HookEntry is the store representation of a lifecycle hook.
-type HookEntry struct {
-	ContainerName string   `json:"container_name"`
-	Phase         string   `json:"phase"`
-	Command       []string `json:"command"`
-	Timeout       int      `json:"timeout"`
-}
-
-// ListHooks returns all hooks for a container.
-func (s *Store) ListHooks(containerName string) ([]HookEntry, error) {
-	var entries []HookEntry
-	prefix := []byte(containerName + "::")
-
-	err := s.db.View(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketHooks)
-		c := b.Cursor()
-		for k, v := c.Seek(prefix); k != nil && bytes.HasPrefix(k, prefix); k, v = c.Next() {
-			var entry HookEntry
-			if err := json.Unmarshal(v, &entry); err != nil {
-				continue
-			}
-			entries = append(entries, entry)
-		}
-		return nil
-	})
-	return entries, err
-}
-
-// SaveHook saves or updates a hook for a container.
-func (s *Store) SaveHook(hook HookEntry) error {
-	data, err := json.Marshal(hook)
-	if err != nil {
-		return fmt.Errorf("marshal hook: %w", err)
-	}
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketHooks)
-		key := []byte(hook.ContainerName + "::" + hook.Phase)
-		return b.Put(key, data)
-	})
-}
-
-// DeleteHook removes a hook for a container.
-func (s *Store) DeleteHook(containerName, phase string) error {
-	return s.db.Update(func(tx *bolt.Tx) error {
-		b := tx.Bucket(bucketHooks)
-		key := []byte(containerName + "::" + phase)
-		return b.Delete(key)
-	})
 }
